fix: Simon Says discards first click
Hypixel seems to have sped up the device. One of their changes seems to cause a button reset after the first sea lantern is placed. This commit changes our detection to use when the sea lantern is changed back to obsidian, instead of when a block is changed to a sea lantern. This seems to happen after the buttons get removed, so it should fix the issue.